+++ python/lancedb/embeddings/cohere.py
@@ -16,7 +16,8 @@ from typing import ClassVar, List, Union
 
 import numpy as np
 
-from .functions import TextEmbeddingFunction, register
+from .registry import register
+from .base import TextEmbeddingFunction
 from .utils import api_key_not_found_help
